Lift document comparison to test utilities
